Initial commit
Initial implementation of `alachemy`, a simple config manager for
Alacritty written for a need of a simple utility to manage alacritty
configuration on the fly.

`alachemy` has a builtin schema to validate any configuration set to
alacritty config. The schema uses a set of regex to match certain values
and value compositions.

Add support to set configuration values for alacritty with a schema that
the passed values are validated against. Currently alacritty
configuration is expected to be found directly under $HOME folder.

NOTE: Some configurations are not supported to be modified through `alachemy`.